프로젝트 구성 에디터용(Debug, Release) 추가 및 ifdef 설정, 에디터에서 Lane 설정 및 Opacity 설정 가능하게 변경
