+++ Client/imgui.ini
@@ -3,6 +3,22 @@ Pos=60,60
 Size=400,400
 
 [Window][Scene Editor]
-Pos=998,398
-Size=121,48
+Pos=100,100
+Size=300,600
+
+[Window][Hierarchy]
+Pos=311,440
+Size=388,302
+
+[Window][Inspector]
+Pos=1387,582
+Size=326,350
+
+[Window][Scene List]
+Pos=1209,98
+Size=337,235
+
+[Window][CameraControl]
+Pos=60,60
+Size=486,289
