+++ Engine/Editor.h
@@ -1,18 +1,20 @@
 #pragma once
-
-class Scene;
+#include "SceneManager.h"
 
 class Editor
 {
 public:
-	void SetScene(Scene* scene) { m_CurrentScene = scene; }
+	Editor(SceneManager& sceneManager) : m_SceneManager(sceneManager) {}
 	void Update();
 private:
-	void DrawHierarchy();
-	void DrawInspector();
+	void DrawSceneList();
+	void DrawGameObjectHierarchy(std::shared_ptr<Scene> currentScene);
+	void DrawGameObjectInspector(std::shared_ptr<Scene> currentScene);
+	void SaveAndLoadFile(std::shared_ptr<Scene> currentScene);
+	void CameraControl(std::shared_ptr<Scene> currentScene);
 
 private:
-	Scene* m_CurrentScene;
+	SceneManager& m_SceneManager;;
 	std::string m_SelectedKey;
 	int m_SelectedIndex = -1;
 };
